+++ canaille/oidc/consents.py
@@ -96,7 +96,7 @@ def restore(user, consent_id):
     else:
         consent.restore()
         if not consent.issue_date:
-            consent.issue_date = datetime.datetime.now()
+            consent.issue_date = datetime.datetime.now(datetime.timezone.utc)
         consent.save()
         flash(_("The access has been restored"), "success")